delete JarvisRunnerUtil ( one used method moved to JarvicUtil )
